Split out doc example from component
